requirements.txt
apscheduler==3.10.4
beautifulsoup4==4.12.3
requests=2.32.2
python-dotenv=1.0.0